requirements.txt
Django~=5.1
djangorestframework~=3.15.2
python-dotenv~=1.0.1
djangorestframework-simplejwt~=5.3.1
djoser~=2.2.3
django-sslserver~=0.22
requests-oauthlib~=2.0.0
environ~=1.0
